Add markdown subtitles in code lines.
Each of code cells performs actions which are needed to conduct the
experiment. I think that ipython notebook should contain code cells with
appropriate markdown explanations so that a beginner can understand the
meaning of the code cells.
That's why I have decided to add markdown subtitles in code lines.